Add X-Bookalope-Error field to the response header.
Adding a X-Bookalope-Error field to the response header allows a client to handle
an error description more conveniently than unpacking the JSON body. Me thinks :)
I added this because the Bookalope Drop client uses client-side Javascript and
requests a Blob XHR response; in case of an error it's not possible to extract
the JSON error from the Blob easily.