Breaking change: sync web and REST API URL formats.
Addressing a bookflow through the web browser does not use the bookflow's book
id as part of the URL's path name, whereas that book id was required as part of
the REST view's path name. This led to two different ways of addressing the same
bookflow. This change shortens the REST API's path name and removes the book id
from it, so that both views have the same path names for the same bookflow.