Minor change to the Scratchpad API: `null` value removes item.
This is a minor but breaking change to the Scratchpad API. Until now
removing individual key-value pairs from a Bookflow’s Scratchpad was
unsupported, and this change adds that feature: a `null` value now
removes an existing key-value pair instead of storing `null` in the
Scratchpad.

I think this makes sense since we don’t have the need to store `null`
values for Bookflows.